requirements.txt
Django==2.1.7
django-rest-auth==0.9.3
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.1.0
Pygments==2.3.1
PyJWT==1.7.1
pytz==2018.9
six==1.12.0
virtualenv==16.4.3
